+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters11/in/A.py
@@ -1,12 +1,12 @@
+import custom
 import tensorflow as tf
 
 
-@tf.function(input_signature=(tf.TensorSpec(shape=[None], dtype=tf.float32),), autograph=False)
-def func(x):
-  print('Tracing with', x)
-  return x
+@custom.decorator(input_signature=None)
+@tf.function(autograph=False)
+def func():
+    pass
 
 
 if __name__ == '__main__':
-    number = tf.constant([1.0, 1.0])
-    func(number)
+    func()
