+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters12/in/requirements.txt
@@ -1,0 +1,1 @@
+tensorflow==2.9.3
