--- edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters13/in/A.py
@@ -1,12 +1,0 @@
-import custom
-import tensorflow as tf
-
-
-@custom.decorator(input_signature=None)
-@tf.function(autograph=False)
-def func():
-    pass
-
-
-if __name__ == '__main__':
-    func()
