--- edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters13/in/requirements.txt
@@ -1,1 +1,0 @@
-tensorflow==2.9.3
