--- edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters14/in/A.py
@@ -1,11 +1,0 @@
-import tensorflow as tf
-
-
-@tf.function(autograph=False)
-@tf.function(jit_compile=True)
-def func(x):
-  return x
-
-
-if __name__ == '__main__':
-    func(tf.constant(1))
