+++ edu.cuny.hunter.hybridize.tests/resources/HybridizeFunction/testComputeParameters9/in/A.py
@@ -1,10 +1,12 @@
 import tensorflow as tf
 
 
-@tf.function(experimental_relax_shapes=True)
-def func():
-    print("Testing")
+@tf.function(input_signature=(tf.TensorSpec(shape=[None], dtype=tf.float32),), autograph=False)
+def func(x):
+  print('Tracing with', x)
+  return x
 
 
 if __name__ == '__main__':
-    func()
+    number = tf.constant([1.0, 1.0])
+    func(number)
