+++ edu.cuny.hunter.hybridize.tests/test cases/edu/cuny/hunter/hybridize/tests/HybridizeFunctionRefactoringTest.java
@@ -675,7 +675,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 
 		assertTrue(!args.hasFuncParam() && !args.hasInputSignatureParam() & !args.hasAutoGraphParam() && args.hasJitCompileParam()
 				&& !args.hasReduceRetracingParam() && !args.hasExperimentalImplementsParam() && !args.hasExperimentalAutographOptParam()
-				&& !args.hasExperimentalRelaxShapesParam() && !args.hasExperimentalCompileParam()
+				&& !args.hasExperimentalRelaxShapesParam() && args.hasExperimentalCompileParam()
 				&& !args.hasExperimentalFollowTypeHintsParam());
 	}
 
@@ -695,7 +695,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 
 		assertTrue(!args.hasFuncParam() && !args.hasInputSignatureParam() & !args.hasAutoGraphParam() && !args.hasJitCompileParam()
 				&& args.hasReduceRetracingParam() && !args.hasExperimentalImplementsParam() && !args.hasExperimentalAutographOptParam()
-				&& !args.hasExperimentalRelaxShapesParam() && !args.hasExperimentalCompileParam()
+				&& args.hasExperimentalRelaxShapesParam() && !args.hasExperimentalCompileParam()
 				&& !args.hasExperimentalFollowTypeHintsParam());
 	}
 
@@ -722,54 +722,10 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 	}
 
 	/**
-	 * Test for #30. This simply tests whether we can parse the tf.function argument experimental_compile.
-	 */
-	@Test
-	public void testComputeParameters8() throws Exception {
-		Set<Function> functions = this.getFunctions();
-		assertNotNull(functions);
-		assertEquals(1, functions.size());
-		Function function = functions.iterator().next();
-		assertNotNull(function);
-
-		assertTrue(function.isHybrid());
-
-		Function.HybridizationParameters args = function.getHybridizationParameters();
-		assertNotNull(args);
-
-		assertTrue(!args.hasFuncParam() && !args.hasInputSignatureParam() & !args.hasAutoGraphParam() && !args.hasJitCompileParam()
-				&& !args.hasReduceRetracingParam() && !args.hasExperimentalImplementsParam() && !args.hasExperimentalAutographOptParam()
-				&& !args.hasExperimentalRelaxShapesParam() && args.hasExperimentalCompileParam()
-				&& !args.hasExperimentalFollowTypeHintsParam());
-	}
-
-	/**
-	 * Test for #30. This simply tests whether we can parse the tf.function argument experimental_relax_shapes.
-	 */
-	@Test
-	public void testComputeParameters9() throws Exception {
-		Set<Function> functions = this.getFunctions();
-		assertNotNull(functions);
-		assertEquals(1, functions.size());
-		Function function = functions.iterator().next();
-		assertNotNull(function);
-
-		assertTrue(function.isHybrid());
-
-		Function.HybridizationParameters args = function.getHybridizationParameters();
-		assertNotNull(args);
-
-		assertTrue(!args.hasFuncParam() && !args.hasInputSignatureParam() & !args.hasAutoGraphParam() && !args.hasJitCompileParam()
-				&& !args.hasReduceRetracingParam() && !args.hasExperimentalImplementsParam() && !args.hasExperimentalAutographOptParam()
-				&& args.hasExperimentalRelaxShapesParam() && !args.hasExperimentalCompileParam()
-				&& !args.hasExperimentalFollowTypeHintsParam());
-	}
-
-	/**
 	 * Test for #30. This simply tests whether we can identify when there are no tf.function args.
 	 */
 	@Test
-	public void testComputeParameters10() throws Exception {
+	public void testComputeParameters8() throws Exception {
 		Set<Function> functions = this.getFunctions();
 		assertNotNull(functions);
 		assertEquals(1, functions.size());
@@ -791,7 +747,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 	 * Test for #30. This simply tests whether we can parse tf.function arguments when we have multiple.
 	 */
 	@Test
-	public void testComputeParameters11() throws Exception {
+	public void testComputeParameters9() throws Exception {
 		Set<Function> functions = this.getFunctions();
 		assertNotNull(functions);
 		assertEquals(1, functions.size());
@@ -813,7 +769,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 	 * Test for #30. Test custom decorator with the same parameter names as tf.function.
 	 */
 	@Test
-	public void testComputeParameters12() throws Exception {
+	public void testComputeParameters10() throws Exception {
 		Set<Function> functions = this.getFunctions();
 		assertNotNull(functions);
 		assertEquals(1, functions.size());
@@ -837,7 +793,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 	 * count the parameters from the tf.function decorator.
 	 */
 	@Test
-	public void testComputeParameters13() throws Exception {
+	public void testComputeParameters11() throws Exception {
 		Set<Function> functions = this.getFunctions();
 		assertNotNull(functions);
 		assertEquals(1, functions.size());
@@ -862,7 +818,7 @@ public class HybridizeFunctionRefactoringTest extends RefactoringTest {
 	 * Test for #30. Tests two different tf.functions. Should only count the parameters of the last one.
 	 */
 	@Test
-	public void testComputeParameters14() throws Exception {
+	public void testComputeParameters12() throws Exception {
 		Set<Function> functions = this.getFunctions();
 		assertNotNull(functions);
 		assertEquals(1, functions.size());
